Add net_stack.rs

+++ src/net_stack.rs
@@ -383,7 +383,7 @@ pub fn epoll_wait(epfd: usize, events: usize, maxevents: usize, timeout: usize) 
             }
         }
         // 超时检查
-        if crate::timer::ticks().wrapping_sub(start_ticks) >= max_ticks {
+        if crate::timer::ticks().wrapping_sub(start_ticks) >= max_ticks as u64 {
             return 0;
         }
     }
